requirements.txt
flet
fpdf
requests
gTTS
Pillow
pystray